+++ changes.md
@@ -2,4 +2,4 @@
 
 * leaflet-elevation 1.7.6
 * collapse control top right always false (not working right)
-* elevation waypoints
+* customize elevation waypoints